move sea result set into backend/sea package
